+++ src/pymonzo/exceptions.py
@@ -11,5 +11,11 @@ class PyMonzoException(Exception):
 
 
 class MonzoAPIException(PyMonzoException):
+    """Monzo API response exception"""
+    pass
+
+
+class UnableToRefreshTokenException(MonzoAPIException):
     """Base Monzo API response exception"""
     pass
+